SB calculation with PHAT ACS data
Run hst/phatsb.py as a script to process one brick at a time, writing a
file with SB in each field.

Note that the ugri uncertainties are apparently huge! They were
calculated using Sage; but something must be wrong. The easier
propagation of uncertainties in adding light within the HST bands shows
that uncertainties are miniscule there.